refactor: extract MarkdownTable parser from MarkdownText
